+++ CHANGELOG.md
@@ -4,9 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
 
-
+## [0.1.3] - 2020-01-31
+### Modified
+- Return `Result<Image>` from `Capture::get_X_image()` instead of just `Image` to accound for possible failure.
 
 ## [0.1.2] - 2019-12-04
 ### Added